fix(build): correct AspectRatio enum names + std::max type deduction
Two errors in swapchain.cpp, both in code I added (blame 583b9303):

1. Config::AspectRatio in emulatorConfig.h was defined as {Stretch, Preserve, Crop}, but the consumer (swapchain.cpp switch at line 759) and the launcher (configuration.h:69) both use {Stretch, Fit16x9, Fit4x3, Integer}. I invented the wrong names when I added the emulator-side enum instead of matching the launcher. Fixed enum to {Stretch, Fit16x9, Fit4x3, Integer} so it matches the launcher and resolves the swapchain switch cases.

2. swapchain.cpp:785 std::max(1, std::min(m_extent.width/sw, m_extent.height/sh)) failed template deduction: m_extent.width is uint32_t so std::min returns uint32_t, conflicting with the int literal 1. Wrapped the std::min in static_cast<int> so both args to std::max are int.

+++ src/graphics/presentation/window/swapchain.cpp
@@ -782,7 +782,7 @@ void Swapchain::RecordPresentCommands(CommandBuffer& command, VulkanImage& sourc
 			case Config::AspectRatio::Integer: {
 				// Largest integer scale that still fits, centred. Uses the source
 				// (guest) resolution as the base, so pixel-art stays crisp.
-				const int scale = std::max(1, std::min(m_extent.width / sw, m_extent.height / sh));
+				const int scale = std::max(1, static_cast<int>(std::min(m_extent.width / sw, m_extent.height / sh)));
 				const int fit_w = sw * scale;
 				const int fit_h = sh * scale;
 				dst_x0 = static_cast<int>(m_extent.width / 2u) - fit_w / 2;
